@@ -1,14 +1,14 @@
 apply plugin: 'com.android.application'
 
 android {
-    compileSdkVersion 26
+    compileSdkVersion 32
     defaultConfig {
         applicationId "com.rareventure.gps2"
         minSdkVersion 21
-        targetSdkVersion 26
+        targetSdkVersion 30
         versionCode 91
         versionName "1.1.48"
-        testInstrumentationRunner "android.support.test.runner.AndroidJUnitRunner"
+        testInstrumentationRunner 'androidx.test.runner.AndroidJUnitRunner'
     }
     buildTypes {
         release {
@@ -16,25 +16,27 @@ android {
             proguardFiles getDefaultProguardFile('proguard-android.txt'), 'proguard-rules.pro'
         }
     }
-    lintOptions {
-        abortOnError false
-        checkReleaseBuilds false
-    }
     compileOptions {
         sourceCompatibility JavaVersion.VERSION_1_8
         targetCompatibility JavaVersion.VERSION_1_8
+    }
+    lint {
+        abortOnError false
+        checkReleaseBuilds false
     }
     //buildToolsVersion '27.0.3'
 }
 
 dependencies {
     implementation fileTree(include: ['*.jar'], dir: 'libs')
-    implementation 'com.android.support:appcompat-v7:26.1.0'
-    implementation 'com.android.support:support-compat:26.1.0'
-    implementation 'ch.acra:acra:4.9.1'
-    testImplementation 'junit:junit:4.12'
-    androidTestImplementation 'com.android.support.test:runner:1.0.1'
-    androidTestImplementation 'com.android.support.test.espresso:espresso-core:3.0.1'
+    implementation 'androidx.appcompat:appcompat:1.4.1'
+    implementation 'androidx.core:core:1.7.0'
+    implementation 'ch.acra:acra:4.11.1'
+    implementation files('libs/bcprov-jdk18on-171.jar')
+    testImplementation 'junit:junit:4.13.2'
+    implementation 'junit:junit:4.13.2'
+    androidTestImplementation 'androidx.test.ext:junit:1.1.3'
+    androidTestImplementation 'androidx.test.espresso:espresso-core:3.4.0'
     // co: This causes a strange problem where I get:
     // java.lang.SecurityException: Neither user nor current process has android.permission WAKE_LOCK
     // I tried asking for the permission, but it causes an infinite loop
@@ -47,4 +49,7 @@ dependencies {
     implementation 'com.mapzen.tangram:tangram:0.10.1'
     implementation 'com.github.tajchert:nammu:1.2.0'
     //this is for asking the user for permissions
+    implementation 'joda-time:joda-time:2.10.14'
+    //implementation 'de.idyl:winzipaes:1.0.1'
+    implementation 'com.codeslap:android-dateslider:0.0.5'
 }
