@@ -25,7 +25,6 @@ import java.io.IOException;
 import com.rareventure.android.AndroidPreferenceSet.AndroidPreferences;
 import com.rareventure.gps2.GTG;
 import com.rareventure.gps2.GTG.GTGEvent;
-import com.rareventure.gps2.GpsTrailerService;
 
 import android.Manifest;
 import android.content.Context;
@@ -36,9 +35,8 @@ import android.location.LocationListener;
 import android.location.LocationManager;
 import android.os.Bundle;
 import android.os.Looper;
-import android.os.SystemClock;
-import android.support.v4.app.ActivityCompat;
-import android.util.Log;
+
+import androidx.core.app.ActivityCompat;
 
 public class GpsReader implements DataReader
 {
