@@ -19,15 +19,9 @@
 */
 package com.rareventure.gps2;
 
-import com.rareventure.android.Util;
-
-import android.app.Activity;
-import android.content.ComponentName;
-import android.content.Context;
 import android.content.Intent;
 import android.os.Bundle;
-import android.support.v4.app.FragmentActivity;
-import android.util.Log;
+import androidx.fragment.app.FragmentActivity;
 
 import pl.tajchert.nammu.Nammu;
 
