@@ -19,7 +19,7 @@
 */
 package com.rareventure.gps2.reviewer.map;
 
-import junit.framework.Assert;
+//import junit.framework.Assert;
 import android.content.Context;
 import android.graphics.Canvas;
 import android.graphics.Paint.FontMetrics;
@@ -28,7 +28,6 @@ import android.util.AttributeSet;
 import android.view.View;
 
 import com.rareventure.android.Util;
-import com.rareventure.gps2.GTG;
 import com.rareventure.gps2.database.TAssert;
 
 /**
